Move RemoteFeedItem to separate file
