+++ FeedLoader/FeedAPI/FeedItemMapper.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-struct RemoteFeedItem: Equatable, Decodable {
-    let id: UUID
-    let description: String?
-    let location: String?
-    let image: URL
-}
-
 final class FeedItemsMapper {
     
     private struct Root: Decodable {
